+++ R/fg_get_bq_config.R
@@ -1,11 +1,10 @@
 
-#' Get CDM Configuration
+#' Get BQ Configuration
 #'
-#' This function retrieves the configuration settings to work with the OMOP-CDM database for a specified environment and data freeze number.
+#' This function retrieves the configuration settings to work with the FinnGen BQ database for a specified environment and data freeze number.
 #'
 #' @param environment Environment must be 'sandbox-' followed by the sandbox number (Alternatively, it can be 'atlasDevelopment' for testing outside sandbox).
 #' @param dataFreezeNumber The data freeze number to retrieve configuration settings for.
-#' @param cohortTableName If a cohort table is needed, it can be set here, (default is "test_cohort_table").
 #' @param atlasDevelopment_gckey If environment = 'atlasDevelopment', the path to the Google Cloud key file.
 #' @param atlasDevelopment_pathToDriver If environment = 'atlasDevelopment', the path to the BigQuery driver.
 #' @param asYaml Whether to return the configuration settings as a YAML string (default is FALSE).
@@ -13,17 +12,16 @@
 #' @return A list of configuration settings for the specified environment and data freeze number.
 #'
 #' @examples
-#' fg_get_cdm_config("sandbox-6", 13)
+#' fg_get_bq_config("sandbox-6", 13)
 #'
 #' @importFrom checkmate assert_string assert_number
 #' @importFrom stringr str_detect str_replace_all
 #' @importFrom yaml yaml.load
 #'
 #' @export
-fg_get_cdm_config <- function(
+fg_get_bq_config <- function(
     environment,
     dataFreezeNumber,
-    cohortTableName = "test_cohort_table",
     atlasDevelopment_gckey =  Sys.getenv("GCP_SERVICE_KEY"),
     atlasDevelopment_pathToDriver = paste0(Sys.getenv("DATABASECONNECTOR_JAR_FOLDER"),"/bigquery/"),
     asYaml = FALSE
@@ -39,7 +37,6 @@ fg_get_cdm_config <- function(
   }
 
   checkmate::assert_number(dataFreezeNumber)
-  checkmate::assert_string(cohortTableName)
 
 
   if (environment == "atlasDevelopment") {
@@ -54,16 +51,11 @@ fg_get_cdm_config <- function(
             pathToDriver: <atlasDevelopment_pathToDriver>
           tempEmulationSchema: atlas-development-270609.sandbox #needed for creating tmp table in BigQuery
           useBigrqueryUpload: true # option for HadesExtras
-        cdm:
-          cdmDatabaseSchema: atlas-development-270609.finngen_omop_r<dataFreezeNumber>
-          vocabularyDatabaseSchema: atlas-development-270609.finngen_omop_r<dataFreezeNumber>
-        cohortTable:
-          cohortDatabaseSchema: atlas-development-270609.sandbox
-          cohortTableName: <cohortTableName>
-        webAPIurl: https://localhost:8080/WebAPI/
+        schemas:
+          sandboxToolsSchema: atlas-development-270609.sandbox_tools_r<dataFreezeNumber>
+          medicalCodesSchema: atlas-development-270609.medical_codes
     ' |>
       stringr::str_replace_all("<dataFreezeNumber>", as.character(dataFreezeNumber)) |>
-      stringr::str_replace_all("<cohortTableName>", cohortTableName) |>
       stringr::str_replace_all("<atlasDevelopment_gckey>", atlasDevelopment_gckey) |>
       stringr::str_replace_all("<atlasDevelopment_pathToDriver>", atlasDevelopment_pathToDriver)
   }
@@ -80,17 +72,12 @@ fg_get_cdm_config <- function(
             pathToDriver: /home/ivm/.jdbc_drivers/bigquery
           tempEmulationSchema: fg-production-<environment>.sandbox #needed for creating tmp table in BigQuery
           useBigrqueryUpload: true # option for HadesExtras
-        cdm:
-          cdmDatabaseSchema: finngen-production-library.finngen_omop_r<dataFreezeNumber>
-          vocabularyDatabaseSchema: finngen-production-library.finngen_omop_r<dataFreezeNumber>
-        cohortTable:
-          cohortDatabaseSchema: fg-production-<environment>.sandbox
-          cohortTableName: <cohortTableName>
-        webAPIurl: https://ohdsi-webapi.app.finngen.fi/WebAPI/
+        schemas:
+          sandboxToolsSchema: atlas-development-270609.sandbox_tools_r<dataFreezeNumber>
+          medicalCodesSchema: atlas-development-270609.medical_codes
     ' |>
       stringr::str_replace_all("<environment>", environment) |>
-      stringr::str_replace_all("<dataFreezeNumber>", as.character(dataFreezeNumber)) |>
-      stringr::str_replace_all("<cohortTableName>", cohortTableName)
+      stringr::str_replace_all("<dataFreezeNumber>", as.character(dataFreezeNumber))
   }
 
   configList <- yaml::yaml.load(configYalm)
@@ -102,6 +89,5 @@ fg_get_cdm_config <- function(
     return(configYalm)
   }
   return(configList)
-
 
 }