+++ dev/03_deploy.R
@@ -17,14 +17,18 @@ browseURL("vignettes/tutorial_add_info.html")
 knitr::knit("vignettes/tutorial_using_omop_cdm.Rmd.orig", output = "vignettes/tutorial_using_omop_cdm.Rmd")
 devtools::build_rmd("vignettes/tutorial_using_omop_cdm.Rmd")
 browseURL("vignettes/tutorial_using_omop_cdm.html")
+
+
+knitr::knit("vignettes/tutorial_connection_handler.Rmd.orig", output = "vignettes/tutorial_connection_handler.Rmd")
+devtools::build_rmd("vignettes/tutorial_connection_handler.Rmd")
+browseURL("vignettes/tutorial_connection_handler.html")
 # increase version
 
 
 
-
-pkgdown::build_site(new_process = FALSE)
-
 #rhub::check_for_cran()
 usethis::use_version()
+
+pkgdown::build_site(new_process = FALSE)
 
 gert::git_push()